Add dna reference assertions
