+++ Assets/Scripts/AI/Evolution/Dna.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d316a0709554c5247a9268e8c1de3fda
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
