+++ Assets/Tests/EditMode/DnaTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9be8c5e47423c1d4fbfeadabaafbc229
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
